+++ packages/webview_flutter/webview_flutter_android/lib/src/android_proxy.dart
@@ -48,6 +48,9 @@ class AndroidWebViewProxy {
         onGeolocationPermissionsShowPrompt,
     void Function(android_webview.WebChromeClient instance)?
         onGeolocationPermissionsHidePrompt,
+    void Function(android_webview.WebChromeClient instance,
+            android_webview.ConsoleMessage message)?
+        onConsoleMessage,
     void Function(
             android_webview.WebChromeClient instance,
             android_webview.View view,