+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -13,7 +13,8 @@ import 'android_webview.dart';
 import 'android_webview.g.dart';
 import 'instance_manager.dart';
 
-export 'android_webview.g.dart' show FileChooserMode;
+export 'android_webview.g.dart'
+    show ConsoleMessage, ConsoleMessageLevel, FileChooserMode;
 
 /// Converts [WebResourceRequestData] to [WebResourceRequest]
 WebResourceRequest _toWebResourceRequest(WebResourceRequestData data) {
@@ -566,6 +567,11 @@ class WebSettingsHostApiImpl extends WebSettingsHostApi {
       enabled,
     );
   }
+
+  /// Helper method to convert instances ids to objects.
+  Future<String> getUserAgentStringFromInstance(WebSettings instance) {
+    return getUserAgentString(instanceManager.getIdentifier(instance)!);
+  }
 }
 
 /// Host api implementation for [JavaScriptChannel].
@@ -906,6 +912,17 @@ class WebChromeClientHostApiImpl extends WebChromeClientHostApi {
       value,
     );
   }
+
+  /// Helper method to convert instances ids to objects.
+  Future<void> setSynchronousReturnValueForOnConsoleMessageFromInstance(
+    WebChromeClient instance,
+    bool value,
+  ) {
+    return setSynchronousReturnValueForOnConsoleMessage(
+      instanceManager.getIdentifier(instance)!,
+      value,
+    );
+  }
 }
 
 /// Flutter api implementation for [DownloadListener].
@@ -1030,6 +1047,13 @@ class WebChromeClientFlutterApiImpl extends WebChromeClientFlutterApi {
         instance,
       );
     }
+  }
+
+  @override
+  void onConsoleMessage(int instanceId, ConsoleMessage message) {
+    final WebChromeClient instance =
+        instanceManager.getInstanceWithWeakReference(instanceId)!;
+    instance.onConsoleMessage?.call(instance, message);
   }
 }