+++ packages/webview_flutter/webview_flutter_wkwebview/pigeons/web_kit.dart
@@ -692,7 +692,7 @@ abstract class WKWebViewHostApi {
   @ObjCSelector('setAllowsBackForwardForWebViewWithIdentifier:isAllowed:')
   void setAllowsBackForwardNavigationGestures(int identifier, bool allow);
 
-  @ObjCSelector('setUserAgentForWebViewWithIdentifier:userAgent:')
+  @ObjCSelector('setCustomUserAgentForWebViewWithIdentifier:userAgent:')
   void setCustomUserAgent(int identifier, String? userAgent);
 
   @ObjCSelector('evaluateJavaScriptForWebViewWithIdentifier:javaScriptString:')
@@ -701,6 +701,9 @@ abstract class WKWebViewHostApi {
 
   @ObjCSelector('setInspectableForWebViewWithIdentifier:inspectable:')
   void setInspectable(int identifier, bool inspectable);
+
+  @ObjCSelector('customUserAgentForWebViewWithIdentifier:')
+  String? getCustomUserAgent(int identifier);
 }
 
 /// Mirror of WKUIDelegate.